remove voting contracts limit
The only usage of this limit is skipping the pruning iteration for
`n` new contracts.
First of all, for the first `n` voting contracts pruning is not
that expensive, as the list is relatively short.
Secondly, if the pruning is skipped at any time, the list can reach
the limit more quickly. And because there is no mechanism to shrink
the voting list, skipping the pruning for first `n` contracts can
make the operation more expensive in the long run.