add numberOfOptions function to VC, change require for VC cstr
Unfortunately, `now` does not mean method invokement time, but
it's equal to the latest block's timestamp. That means a user
won't be able to see the results after `votingEndTime` if the
latest block has a smaller timestamp.
We assume that with a well-working Ethereum infrastructure
the timestamps will update with a 20s interval.